+++ apps/explorer/lib/explorer/chain/block.ex
@@ -81,7 +81,8 @@ defmodule Explorer.Chain.Block do
           round: non_neg_integer(),
           extra_data: Data.t(),
           online: %Ecto.Association.NotLoaded{} | boolean(),
-          base_fee_per_gas: Wei.t()
+          base_fee_per_gas: Wei.t(),
+          is_empty: boolean()
         }
 
   @primary_key {:hash, Hash.Full, autogenerate: false}
@@ -100,6 +101,7 @@ defmodule Explorer.Chain.Block do
     field(:round, :integer)
     field(:extra_data, Data)
     field(:base_fee_per_gas, Wei)
+    field(:is_empty, :boolean)
 
     timestamps()