+++ apps/explorer/priv/repo/migrations/20211018072347_add_is_empty_index.exs
@@ -1,0 +1,17 @@
+defmodule Explorer.Repo.Migrations.AddIsEmptyIndex do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+  @disable_migration_lock true
+
+  def change do
+    create(
+      index(
+        :blocks,
+        ~w(consensus)a,
+        name: :empty_consensus_blocks,
+        where: "is_empty IS NULL",
+        concurrently: true
+      )
+    )
+  end
+end